Make sure loading works correctly on user dashboard
