+++ app/components/Jobs/CreateJobApplication.js
@@ -7,9 +7,11 @@ import viewDictionary from './viewDictionary';
 const CreateJobApplication = ({ view, ...restProps }) => {
   useEffect(() => {
     window.scrollTo(0, 0);
-    document.title = 'Jobs';
+    document.title = 'Start Job Application';
   }, []);
+
   const ViewToRender = viewDictionary[view];
+
   return (
     <JobsContainer>
       <ViewToRender {...restProps} />
